Moved sql_insert_array and sql_update_array to a new location and name. Now in db_utils.php and called db_insert_array and db_update_array.
Created GuiContainer.php which is intended to be extended when creating a guicontrols with embedded (child) guiControls.

It still has a couple bugs, but is functional enough for usage. The major bug I have found is that it validates one time later, so it gives you the value and message for the value you typed last time.