Fixed major bug in guiControls - Zoop no longer pushes js or css pages to the browser. removed sUrls variable usage. Created a new css library called messages for instant formatting of important messages. Created new smarty plugin to make including js zoop_files very easy.
Reviesed all guiControls to work with new js validation library.
Fixed bugs in the captcha guicontrol.
Made a few other changes to the guiControl object and employed new changes in the individual controls.
Created a new control openwysiwyg for the openwysiwyg js library.
Added support for new js validation library to the validate component.
Fixed bugs in zoopfile it now displays the proper mimetype.. the headers were displayed in the wrong order before.

** Made slight change in the API for guicontrols **. 
	Now individual controls render() call only draws their immediate control html. 
	A new call renderControl() is used by guicontrols to produce the viewstate and error messages
	Makes creating and implementing guicontrols much simplier and standardized.
	The smarty plugin guicontrols calls the new renderControl instead of the previously used render()